Refactor/#47: TripJoinResponse클래스 위치 변경
