+++ src/main/java/gdg/travodobackend/app/travel/controller/TripController.java
@@ -1,6 +1,6 @@
 package gdg.travodobackend.app.travel.controller;
 
-import gdg.travodobackend.app.travel.dto.TripInviteCodeResponse;
+import gdg.travodobackend.app.travel.dto.trip.TripInviteCodeResponse;
 import gdg.travodobackend.app.travel.dto.member.TripMemberResponse;
 import gdg.travodobackend.app.travel.dto.trip.*;
 import gdg.travodobackend.app.travel.service.TripService;
