+++ src/main/java/gdg/travodobackend/app/travel/service/TripService.java
@@ -1,6 +1,6 @@
 package gdg.travodobackend.app.travel.service;
 
-import gdg.travodobackend.app.travel.dto.TripInviteCodeResponse;
+import gdg.travodobackend.app.travel.dto.trip.TripInviteCodeResponse;
 import gdg.travodobackend.app.travel.dto.member.TripMemberResponse;
 import gdg.travodobackend.app.travel.dto.trip.*;
 import gdg.travodobackend.app.travel.entity.Trip;
